+++ gradle.properties
@@ -1,5 +1,5 @@
 org.gradle.caching=true
-group=org.ballerinalang
+group=io.ballerina.stdlib
 version=1.1.0-alpha5-SNAPSHOT
 ballerinaLangVersion=2.0.0-alpha6-SNAPSHOT
