+++ native/src/main/java/io/ballerina/stdlib/xmldata/utils/Constants.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) 2021 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.ballerina.stdlib.xmldata.utils;
+
+/**
+ * Constants used in Ballerina XmlData library.
+ *
+ * @since 1.1.0
+ */
+public class Constants {
+
+    private Constants() {}
+
+    public static final String OPTIONS_ATTRIBUTE_PREFIX = "attributePrefix";
+    public static final String OPTIONS_ARRAY_ENTRY_TAG = "arrayEntryTag";
+    public static final String OPTIONS_PRESERVE_NS = "preserveNamespaces";
+}
