+++ native/src/main/java/io/ballerina/stdlib/xmldata/utils/XmlDataUtils.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright (c) 2021 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.ballerina.stdlib.xmldata.utils;
+
+import io.ballerina.runtime.api.creators.ErrorCreator;
+import io.ballerina.runtime.api.utils.StringUtils;
+import io.ballerina.runtime.api.values.BError;
+
+/**
+ * A util class for the XmlData package's native implementation.
+ *
+ * @since 1.1.0
+ */
+public class XmlDataUtils {
+
+    private static final String ERROR = "Error";
+
+    public static BError getError(String message) {
+        return ErrorCreator.createDistinctError(ERROR, ModuleUtils.getModule(),
+                StringUtils.fromString(message));
+    }
+}
