--- xmldata-native/build.gradle
@@ -1,72 +1,0 @@
-/*
- * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-plugins {
-    id 'java'
-    id 'checkstyle'
-    id 'com.github.spotbugs'
-}
-
-description = 'Ballerina - Xmlutils Java Utils'
-
-dependencies {
-    checkstyle project(':build-config:checkstyle')
-    checkstyle "com.puppycrawl.tools:checkstyle:${puppycrawlCheckstyleVersion}"
-
-    compile group: 'org.ballerinalang', name: 'ballerina-lang', version: "${ballerinaLangVersion}"
-    compile group: 'org.ballerinalang', name: 'ballerina-runtime', version: "${ballerinaLangVersion}"
-}
-
-checkstyle {
-    toolVersion "${checkstyleToolVersion}"
-    configFile rootProject.file("build-config/checkstyle/build/checkstyle.xml")
-    configProperties = ["suppressionFile" : file("${rootDir}/build-config/checkstyle/build/suppressions.xml")]
-}
-
-checkstyleMain.dependsOn(":build-config:checkstyle:downloadMultipleFiles")
-
-def excludePattern = '**/module-info.java'
-tasks.withType(Checkstyle) {
-    exclude excludePattern
-}
-
-spotbugsMain {
-    effort "max"
-    reportLevel "low"
-    reportsDir = file("$project.buildDir/reports/spotbugs")
-    reports {
-        html.enabled true
-        text.enabled = true
-    }
-    def excludeFile = file("${rootDir}/spotbugs-exclude.xml")
-    if(excludeFile.exists()) {
-        excludeFilter = excludeFile
-    }
-}
-
-spotbugsTest {
-    enabled = false
-}
-
-compileJava {
-    doFirst {
-        options.compilerArgs = [
-                '--module-path', classpath.asPath,
-        ]
-        classpath = files()
-    }
-}
